PR feedback

+++ Source/HTTP/WinHttp/winhttp_proxy_helper.cpp
@@ -12,7 +12,8 @@ http_internal_wstring WinHttpProvider::BuildNamedProxyString(_In_ const xbox::ht
     }
     if (proxyUri.Port() > 0)
     {
-        auto portStr = std::to_wstring(static_cast<unsigned int>(proxyUri.Port()));
+        // uint16_t promotes safely; explicit cast not required
+        auto portStr = std::to_wstring(proxyUri.Port());
         http_internal_wstring result = wProxyHost;
         result.push_back(L':');
         result.append(portStr.c_str());
